list downloading, infinite scrolling

+++ app/src/main/java/com/marcin/jasi/littleandroidapp/photosList/domain/entity/GetPhotosListParams.kt
@@ -1,3 +1,3 @@
 package com.marcin.jasi.littleandroidapp.photosList.domain.entity
 
-class GetPhotosListParams(var lastId: Long, var amount: Long)
+class GetPhotosListParams(var lastId: Int, var amount: Long)
